3rd - Added Follow/Unfollow, like & comment post and notifications page functionality

+++ client/src/components/common/Post.jsx
@@ -6,6 +6,7 @@ import { BiRepost } from "react-icons/bi";
 import { FaRegComment, FaRegHeart, FaTrash } from "react-icons/fa";
 import { FaRegBookmark } from "react-icons/fa6";
 import { Link } from "react-router-dom";
+import { formatPostDate } from "../../utils/date/index.js";
 import LoadingSpinner from "./LoadSpinner";
 
 const Post = ({ post }) => {
@@ -14,7 +15,14 @@ const Post = ({ post }) => {
     const { data: authUser } = useQuery({ queryKey: ['authUser'] });
     const queryClient = useQueryClient();
 
-    const { mutate: deletePost, isPending } = useMutation({
+    const postOwner = post.user;
+    const isLiked = post.likes.includes(authUser._id);
+
+    const isMyPost = authUser._id === post.user._id;
+
+    const formattedDate = formatPostDate(post.createdAt);
+
+    const { mutate: deletePost, isPending: isDeleting } = useMutation({
         mutationFn: async () => {
             try {
                 const res = await axios.post(`http://localhost:8000/api/posts/${post._id}`);
@@ -31,26 +39,73 @@ const Post = ({ post }) => {
             // Invalidating query to refetch data
             queryClient.invalidateQueries({ queryKey: ['posts'] });
         }
+    });
+
+    const { mutate: likePost, isPending: isLiking } = useMutation({
+        mutationFn: async () => {
+            try {
+                const res = await axios.post(`http://localhost:8000/api/posts/like/${post._id}`);
+                const data = res.data;
+
+                if (res.status !== 200) throw new Error(data.error || "Something Went Wrong!!");
+                return data;
+            } catch (error) {
+                throw new Error(error);
+            }
+        },
+        onSuccess: (updatedLikes) => {
+            // queryClient.invalidateQueries({ queryKey: ['posts'] }); // reloads on liking
+
+            queryClient.setQueryData(['posts'], (oldData) => {
+                return oldData.map(p => {
+                    if (p._id === post._id) {
+                        return { ...p, likes: updatedLikes };
+                    }
+                    return p;
+                });
+            });
+        }
+    });
+
+    const { mutate: commentPost, isPending: isCommenting } = useMutation({
+        mutationFn: async () => {
+            try {
+                const res = await axios.post(`http://localhost:8000/api/posts/comments/${post._id}`, { text: comment });
+                const data = res.data;
+
+                if (res.status !== 200) throw new Error(data.error || "Something Went Wrong!!");
+
+                return data;
+            } catch (error) {
+                throw new Error(error);
+            }
+        },
+        onSuccess: () => {
+            toast.success("Comment Posted Successfully!!");
+            setComment('');
+            queryClient.invalidateQueries({ queryKey: ['posts'] });
+        },
+        onError: (error) => {
+            toast.error(error.message);
+        }
     })
 
-    const postOwner = post.user;
-    const isLiked = false;
-
-    const isMyPost = authUser._id === post.user._id;
-
-    const formattedDate = "1h";
-
-    const isCommenting = false;
-
     const handleDeletePost = () => {
-        deletePost();
+        if (confirm("Delete Post?")) {
+            deletePost();
+        }
     };
 
     const handlePostComment = (e) => {
         e.preventDefault();
+        if (isCommenting) return;
+        commentPost();
     };
 
-    const handleLikePost = () => { };
+    const handleLikePost = () => {
+        if (isLiking) return;
+        likePost();
+    };
 
     return (
         <>
@@ -72,8 +127,8 @@ const Post = ({ post }) => {
                         </span>
                         {isMyPost && (
                             <span className='flex justify-end flex-1'>
-                                {!isPending && <FaTrash className='cursor-pointer text-primary hover:text-red-500' onClick={handleDeletePost} />}
-                                {isPending && (
+                                {!isDeleting && <FaTrash className='cursor-pointer text-primary hover:text-red-500' onClick={handleDeletePost} />}
+                                {isDeleting && (
                                     <LoadingSpinner size="md" />
                                 )}
                             </span>
@@ -143,7 +198,7 @@ const Post = ({ post }) => {
                                         />
                                         <button className='btn btn-primary rounded-full btn-sm text-black px-4'>
                                             {isCommenting ? (
-                                                <span className='loading loading-spinner loading-md'></span>
+                                                <LoadingSpinner size="md" />
                                             ) : (
                                                 "Post"
                                             )}
@@ -159,10 +214,11 @@ const Post = ({ post }) => {
                                 <span className='text-sm text-slate-500 group-hover:text-green-500'>0</span>
                             </div>
                             <div className='flex gap-1 items-center group cursor-pointer' onClick={handleLikePost}>
-                                {!isLiked && (
+                                {isLiking && <LoadingSpinner size="md" />}
+                                {!isLiked && !isLiking && (
                                     <FaRegHeart className='w-4 h-4 cursor-pointer text-slate-500 group-hover:text-pink-500' />
                                 )}
-                                {isLiked && <FaRegHeart className='w-4 h-4 cursor-pointer text-pink-500 ' />}
+                                {isLiked && !isLiking && <FaRegHeart className='w-4 h-4 cursor-pointer text-pink-500 ' />}
 
                                 <span
                                     className={`text-sm text-slate-500 group-hover:text-pink-500 ${isLiked ? "text-pink-500" : ""
